step1 - DateFormatter -> DateConverter
제조일자 출력 형식을 깔끔하게 다듬었다
