+++ VendingMachineApp/VendingMachineApp/Model/Beverage.swift
@@ -33,7 +33,7 @@ class Beverage: CustomStringConvertible {
     }
     
     var description: String {
-        return "\(brand), \(volume)ml, \(price)원, \(name), \(manufacturingDate)"
+        return "\(brand), \(volume)ml, \(price)원, \(name), \(DateConverter.format(from: manufacturingDate))"
     }
 }
